+++ src/nl/hannahsten/texifyidea/lang/LatexPackage.kt
@@ -27,6 +27,7 @@ open class LatexPackage @JvmOverloads constructor(
         // Predefined packages.
         val DEFAULT = LatexPackage("")
 
+        val ACRONYM = LatexPackage("acronym")
         val ADDTOLUATEXPATH = LatexPackage("addtoluatexpath")
         val ALGORITHM2E = LatexPackage("algorithm2e")
         val ALGORITHMICX = LatexPackage("algorithmicx")